Update `std::to_string` casts to use `long long` for pointer conversion

+++ src/GUI/Widgets/LayoutPanel.h
@@ -31,7 +31,7 @@ public:
 class LayoutPanel : public Panel {
 public:
     LayoutPanel() {
-        name = "LayoutPanel_"+std::to_string(reinterpret_cast<long>(this));
+        name = "LayoutPanel_"+std::to_string(reinterpret_cast<long long>(this));
     }
     LinearLayout layout{0,5,5,false};
